+++ src/stable/base/FeeConfiguration.sol
@@ -98,6 +98,6 @@ abstract contract FeeConfiguration is IFeeConfiguration, BlockNumberish {
     /// @param _poolId The pool ID to reset fee state for
     function _resetFeeState(PoolId _poolId) internal {
         feeState[_poolId].previousFeeE12 = uint40(FeeCalculation.UNDEFINED_FLEXIBLE_FEE_E12);
-        feeState[_poolId].blockNumber = _getBlockNumberish();
+        feeState[_poolId].blockNumber = uint40(_getBlockNumberish());
     }
 }
