+++ src/stable/interfaces/IFeeConfiguration.sol
@@ -13,7 +13,7 @@ struct FeeConfig {
 struct FeeState {
     uint40 previousFeeE12; // Last flexible fee charged in 1e12 precision; used for exponential decay calculation
     uint160 previousSqrtAmmPriceX96; // AMM sqrt price at last swap; used to detect price movement direction
-    uint256 blockNumber; // Block when fee was last updated; determines decay based on blocks elapsed
+    uint40 blockNumber; // Block when fee was last updated; determines decay based on blocks elapsed
 }
 
 /// @notice Interface for the FeeConfiguration
